+++ src/libAtomVM/memory.c
@@ -25,6 +25,7 @@
 #include "debug.h"
 #include "dictionary.h"
 #include "erl_nif_priv.h"
+#include "globalcontext.h"
 #include "list.h"
 #include "memory.h"
 #include "refc_binary.h"
@@ -292,7 +293,7 @@ static enum MemoryGCResult memory_gc(Context *ctx, size_t new_size, size_t num_r
 
     ctx->heap.heap_ptr = temp_end;
 
-    memory_sweep_mso_list(old_mso_list, ctx->global);
+    memory_sweep_mso_list(old_mso_list, ctx->global, false);
     ctx->heap.root->mso_list = new_mso_list;
 
     memory_destroy_heap_fragment(old_root_fragment);
@@ -711,8 +712,11 @@ void memory_heap_append_fragment(Heap *heap, HeapFragment *fragment, term mso_li
     }
 }
 
-void memory_sweep_mso_list(term mso_list, GlobalContext *global)
+void memory_sweep_mso_list(term mso_list, GlobalContext *global, bool from_task)
 {
+#ifndef AVM_TASK_DRIVER_ENABLED
+    UNUSED(from_task)
+#endif
     term l = mso_list;
     while (l != term_nil()) {
         term h = term_get_list_head(l);
@@ -723,7 +727,15 @@ void memory_sweep_mso_list(term mso_list, GlobalContext *global)
             // it has been moved, so it is referenced
         } else if (term_is_refc_binary(h) && !term_refc_binary_is_const(h)) {
             // unreferenced binary; decrement reference count
-            refc_binary_decrement_refcount((struct RefcBinary *) term_refc_binary_ptr(h), global);
+#ifdef AVM_TASK_DRIVER_ENABLED
+            if (from_task) {
+                globalcontext_refc_decrement_refcount_from_task(global, (struct RefcBinary *) term_refc_binary_ptr(h));
+            } else {
+#endif
+                refc_binary_decrement_refcount((struct RefcBinary *) term_refc_binary_ptr(h), global);
+#ifdef AVM_TASK_DRIVER_ENABLED
+            }
+#endif
         }
         l = term_get_list_tail(l);
     }