+++ src/platforms/esp32/components/avm_sys/sys.c
@@ -68,14 +68,6 @@
 
 #define TAG "sys"
 
-#ifndef AVM_NO_SMP
-#define SMP_MUTEX_LOCK(mtx) smp_mutex_lock(mtx)
-#define SMP_MUTEX_UNLOCK(mtx) smp_mutex_unlock(mtx)
-#else
-#define SMP_MUTEX_LOCK(mtx)
-#define SMP_MUTEX_UNLOCK(mtx)
-#endif
-
 static Context *port_driver_create_port(const char *port_name, GlobalContext *global, term opts);
 
 static void *select_thread_loop(void *);
@@ -140,7 +132,7 @@ static void receive_events(GlobalContext *glb, TickType_t wait_ticks)
             sender = event_source;
         }
 
-#ifndef AVM_NO_SMP
+#if !defined(AVM_NO_SMP) || defined(AVM_TASK_DRIVER_ENABLED)
         if (sender == CAST_FUNC_TO_VOID_PTR(sys_signal)) {
             // We've been signaled
             return;
@@ -167,13 +159,11 @@ void sys_poll_events(GlobalContext *glb, int timeout_ms)
     receive_events(glb, wait_ticks);
 }
 
-#ifndef AVM_NO_SMP
 void sys_signal(GlobalContext *glb)
 {
     void *queue_item = CAST_FUNC_TO_VOID_PTR(sys_signal);
     xQueueSendToBack(event_queue, &queue_item, 0);
 }
-#endif
 
 void sys_time(struct timespec *t)
 {
@@ -810,11 +800,15 @@ term esp_err_to_term(GlobalContext *glb, esp_err_t status)
 int sys_mbedtls_entropy_func(void *entropy, unsigned char *buf, size_t size)
 {
 #ifndef MBEDTLS_THREADING_C
+#ifndef AVM_NO_SMP
     struct ESP32PlatformData *platform
         = CONTAINER_OF(entropy, struct ESP32PlatformData, entropy_ctx);
     SMP_MUTEX_LOCK(platform->entropy_mutex);
+#endif
     int result = mbedtls_entropy_func(entropy, buf, size);
+#ifndef AVM_NO_SMP
     SMP_MUTEX_UNLOCK(platform->entropy_mutex);
+#endif
 
     return result;
 #else
@@ -844,8 +838,12 @@ mbedtls_entropy_context *sys_mbedtls_get_entropy_context_lock(GlobalContext *glo
 
 void sys_mbedtls_entropy_context_unlock(GlobalContext *global)
 {
+#ifndef AVM_NO_SMP
     struct ESP32PlatformData *platform = global->platform_data;
     SMP_MUTEX_UNLOCK(platform->entropy_mutex);
+#else
+    UNUSED(global);
+#endif
 }
 
 mbedtls_ctr_drbg_context *sys_mbedtls_get_ctr_drbg_context_lock(GlobalContext *global)
@@ -876,6 +874,10 @@ mbedtls_ctr_drbg_context *sys_mbedtls_get_ctr_drbg_context_lock(GlobalContext *g
 
 void sys_mbedtls_ctr_drbg_context_unlock(GlobalContext *global)
 {
+#ifndef AVM_NO_SMP
     struct ESP32PlatformData *platform = global->platform_data;
     SMP_MUTEX_UNLOCK(platform->random_mutex);
+#else
+    UNUSED(global);
+#endif
 }